Settings: one tabbed page shared by the phone and the Cockpit
The two surfaces did not offer the same settings. The Cockpit had Notifications,
AI and Car Mode, with the microphone and transcription checks on a different page
entirely. The phone had a single untabbed "AI settings" scroll: no notification
settings, no Car Mode end phrase, and the two checks linked out to screens of
their own. Same account, two different products.

The page body now lives in client-core and both shells mount it, so there is one
tab set and one copy of every card:

  Notifications - snooze lengths, display time zone, notifications on this device
  AI            - hosted provider, thinking + fast models, spoken language, voice
  Transcription - the transcription model, how your microphones are measuring,
                  Test microphone, Test transcription
  Car Mode      - model, end phrase, live phrase tester

Transcription is a tab because "dictation keeps coming out wrong" is its own
problem and the four things that diagnose it were in four places. The Cockpit's
Transcription Health page keeps the report over time and links across; it no
longer carries a second copy of the checks. The Car Mode model left the phone's
AI screen for the Car Mode tab.

The phone gains snooze, time zone, notifications and the end phrase. Its screen
is titled Settings and is only a frame around the shared body; styles.css re-tunes
the shared cards for touch under a .screen ancestor, so layout differs and content
cannot. ?tab= resolves the same on both, and /mic-test and /transcription-test
redirect into the Transcription tab so existing bookmarks still land on what they
asked for.

Also: neither dev config proxied /voice-quality or the transcription report
endpoints, so those panels could not be exercised under npm run dev. The Cockpit
proxies the three exact /transcription API paths rather than the prefix - its own
Transcription page is routed at /transcription and would otherwise answer its own
URL with JSON.
