+++ packages/client-core/package.json
@@ -10,6 +10,9 @@
     "./dictation/DictationDialog": "./src/dictation/DictationDialog.tsx",
     "./dictation/MicTestPanel": "./src/dictation/MicTestPanel.tsx",
     "./dictation/TranscriptionTestPanel": "./src/dictation/TranscriptionTestPanel.tsx",
+    "./settings/SettingsTabs": "./src/settings/SettingsTabs.tsx",
+    "./settings/NotificationsTab": "./src/settings/NotificationsTab.tsx",
+    "./transcription/MicrophoneQualityPanel": "./src/transcription/MicrophoneQualityPanel.tsx",
     "./*": "./src/*.ts"
   },
   "scripts": {
